gumjs: Migrate to Interceptor.get_live_replacement_invocation()
To simplify things, and also guarantee that the specific replacement
function has a live call.
